requirements.txt
Flask==3.0.3
pdfplumber==0.11.4
pandas==2.2.3
psycopg2-binary==2.9.9
requests==2.32.3